// src/app/Models/GetRoteApiDataClass.ts
// Generated by https://quicktype.io

export interface GetRoteApiDataClass {
    DeptCode:              string;
    DeptName:              string;
    RoteID:                number;
    RoteCode:              string;
    RoteNameC:             string;
    OnDateTime:            string;
    OffDateTime:           string;
    OnTime:                string;
    OffTime:               string;
    WorkHours:             number;
    DWorkHours:            number;
    OnTimeEarliest:        number;
    OffTimeLatest:         number;
    OtBeginTime:           string;
    YearRestHours:         number;
    LeaveOffTime:          string;
    FlexibleMinute:        number;
    FlexibleMinuteForward: number;
    FlexibleMinuteBehind:  number;
    LateMinute:            number;
    WorkInterval:          number;
    IsCard:                boolean;
    IsShift:               boolean;
    IsDifferShift:         boolean;
    Ride:                  boolean;
    Seq:                   number;
    Holiday:               boolean;
    RoteRest:              any[];
    RoteMapping:           RoteMapping;
    uiShowRoteName:        string;
    OnTime_calCrossDay:    boolean;
    OffTime_calCrossDay:   boolean;
    uiWorkTime:{
        Hour:number,
        Min:number
    };
}

export interface RoteMapping {
    RoteMappingCode: string;
    RoteMappingName: string;
    RoteID:          number;
}