// src/app/Models/GetAttendSumRoteGetApiDataClass.ts
// Generated by https://quicktype.io

export interface GetAttendSumRoteGetApiDataClass {
    RoteID: number;
    Rote:   Rote;
    Sum:    number;
}

export interface Rote {
    RoteID:                number;
    RoteCode:              string;
    RoteNameC:             string;
    OnDateTime:            string;
    OffDateTime:           string;
    OnTime:                string;
    OffTime:               string;
    WorkHours:             number;
    DWorkHours:            number;
    OnTimeEarliest:        number;
    OffTimeLatest:         number;
    OtBeginTime:           string;
    YearRestHours:         number;
    LeaveOffTime:          string;
    FlexibleMinute:        number;
    FlexibleMinuteForward: number;
    FlexibleMinuteBehind:  number;
    LateMinute:            number;
    WorkInterval:          number;
    IsCard:                boolean;
    IsShift:               boolean;
    IsDifferShift:         boolean;
    Ride:                  boolean;
    Seq:                   number;
    Holiday:               boolean;
    RoteRest:              RoteREST[];
    RoteMapping:           null;
}

export interface RoteREST {
    RoteID:       number;
    Seq:          number;
    TimeB:        string;
    TimeE:        string;
    Minute:       number;
    IsNormalAbs:  boolean;
    IsNormalOt:   boolean;
    IsHoliDayAbs: boolean;
    IsHoliDayOt:  boolean;
}